requirements.txt

flask
pyjwt
dotenv
pymsql
flask-bcrypt
flask_sqlalchemy